requirements.txt
python-telegram-bot>=20.0,<21.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
pydantic>=2.0.0